+++ teamwork/src/main/java/com/future/teamwork/dao/UserDao.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.future.teamwork.domain.User;
 
 @Repository
-public interface  UserDao  extends JpaRepository<User, Integer> {
+public interface  UserDao  extends JpaRepository<User, Long> {
 	
 	User getByUserNameAndStatus(String userName,int status);
