+++ teamwork/src/main/java/com/future/teamwork/service/SystemLogService.java
@@ -2,7 +2,7 @@ package com.future.teamwork.service;
 
 import com.future.teamwork.domain.SystemLog;
 
-public interface SystemLogService extends BaseService<SystemLog, Integer>{
+public interface SystemLogService extends BaseService<SystemLog, Long>{
