+++ teamwork/src/main/java/com/future/teamwork/service/impl/SystemLogImpl.java
@@ -8,7 +8,7 @@ import com.future.teamwork.domain.SystemLog;
 import com.future.teamwork.service.SystemLogService;
 
 @Service
-public class SystemLogImpl extends BaseServiceImpl<SystemLog, Integer> implements SystemLogService {
+public class SystemLogImpl extends BaseServiceImpl<SystemLog, Long> implements SystemLogService {
 
 	@Autowired
     private SystemLogDao systemLogDao;
